% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEA.R
\name{data_process}
\alias{data_process}
\title{Data process}
\usage{
data_process(Data, group, norm.form = "CPM", is.normalized = FALSE)
}
\arguments{
\item{Data}{single-cell RNA-seq matrix. The format could be raw-counts, FPKM/RPKM, TPM or UMI-counts. The matrix need include gene names and cell names.}

\item{group}{group information. The cell need be divided into two category.}

\item{norm.form}{character item. We provide several normalized method for raw-counts data. The method include "TMM","RLE", "CPM". The default is "CPM".}

\item{is.normalized}{logical. A logical flag to determin whether or not the input dataset normalizes. If TRUE, we will take the Data as normcounts and input for downstream analysis. If not, we provide method for the process.}
}
\value{
\itemize{
  \item \strong{sce} :  A \code{\linkS4class{SingleCellExperiment}} item. The object include expression matrix, group information. The expression matrix contains counts and normcounts.
}
}
\description{
This function focus on dealing various single-cell RNA-seq input and unifying output format. \cr
}
\details{
We take \code{\link[monocle]{relative2abs}} transfering relative expression values into absolute transcript counts.
However, the process maybe break the original dataset statistical properties. Hence, we advise user don't normalize firstly.
}
\examples{
data("Grun.counts.hvg")
data("Grun.group.information")
sce <- data_process(Data = Grun.counts.hvg, group = Grun.group.information)
}
\author{
Huisheng, Li,  <lihs@mails.ccnu.edu.cn>
}
