% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grun.counts.hvg}
\alias{Grun.counts.hvg}
\title{Gene expression count matrix  of Grun}
\format{
a large matrix
}
\usage{
data(Grun.counts.hvg)
}
\description{
We obtain the preprocessing Grun data from Soneson. Two group cell types, WG and YPS, are
selected to perform DE analysis. To demonstrate the effectiveness of the scDEA, We use the function FindVariableFeatures in Seurat R package
to select 2000 highly variable genes. In this study, the Grun data contains 2000 genes, 338 cells of WG and
378 cells of YPS.
}
\examples{
data(Grun.counts.hvg)
}
\keyword{datasets}
