% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grun.counts.matrix}
\alias{Grun.counts.matrix}
\title{Gene expression count matrix  of Grun}
\format{
a large matrix
}
\usage{
data(Grun.counts.matrix)
}
\description{
We obtain the preprocessing Grun data from Soneson. Two group cell types, WG and YPS, are
selected to perform DE analysis.  We use the function FindVariableFeatures in Seurat R package
to select 20000 highly variable genes. In this study, the Grun data contains 20000 genes, 338 cells of WG and
378 cells of YPS.
}
\examples{
data(Grun.counts.matrix)
}
\keyword{datasets}
