% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grun.group.information}
\alias{Grun.group.information}
\title{Cell type lebals of Grun data}
\format{
a vector
}
\usage{
data(Grun.group.information)
}
\description{
Cell type lebals of Grun data
}
\examples{
data(Grun.group.information)
}
\keyword{datasets}
