% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEA.R
\name{lancaster.combination}
\alias{lancaster.combination}
\title{Combining the p-values identified by DE ananlysis methods}
\usage{
lancaster.combination(Pvals, weight = TRUE, trimmed = 0.2)
}
\arguments{
\item{Pvals}{a p-value matrix. The rows represent genes and the columns correspond to the individual DE analysis methods.}

\item{weight}{a boolean variable that defines whether to use spearman correlation measure the similarity between different DE analysis methods. Default is "TRUE".}

\item{trimmed}{a real number between 0 and 0.5 to trim p-values. Default value is 0.2.}
}
\value{
a vector represents the combined p-value for each gene.
}
\description{
This function is used to combine the results of differential expression anlaysis and obtain an uniform result.
}
\author{
Huisheng, Li,  <lihs@mails.ccnu.edu.cn>
}
