% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEA.R
\name{normalized}
\alias{normalized}
\title{Normalized process}
\usage{
normalized(counts.matrix, method = "CPM")
}
\arguments{
\item{counts.matrix}{count expression matrix}

\item{method}{character. "TMM", "RLE", "CPM". The default value is "CPM".}
}
\description{
The function provide several normalized methods
}
