% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEA.R
\name{scDEA.p.adjust}
\alias{scDEA.p.adjust}
\title{adjusting the p-values identified by individual DE ananlysis methods and scDEA}
\usage{
scDEA.p.adjust(combination.Pvals, adjusted.method = "bonferroni")
}
\arguments{
\item{combination.Pvals}{a vector of combined p-value for each gene.}

\item{adjusted.method}{a string variable specifying the type of p.adjust method used in t-test method.
Possible values: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default is "bonferroni".}
}
\value{
adjusted p-value ,
}
\description{
This function is used to adjust the p-values generated by individual DE analysis methods and scDEA.
}
