% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEA.R
\name{scDEA_individual_methods}
\alias{scDEA_individual_methods}
\title{Run individual DE analysis methods to perform DE analysis on scRNA-seq datasets.}
\usage{
scDEA_individual_methods(
  raw.count,
  cell.label,
  is.normalized = FALSE,
  verbose = TRUE,
  BPSC = TRUE,
  DEsingle = TRUE,
  DESeq2 = TRUE,
  edgeR = TRUE,
  MAST = TRUE,
  monocle = TRUE,
  scDD = TRUE,
  Ttest = TRUE,
  Wilcoxon = TRUE,
  limma = TRUE,
  Seurat = TRUE,
  zingeR.edgeR = TRUE,
  BPSC.coef = 2,
  BPSC.normalize = "CPM",
  BPSC.parallel = TRUE,
  DEsingle.parallel = TRUE,
  DEsingle.normalize = "CPM",
  DESeq2.test = "LRT",
  DESeq2.parallel = TRUE,
  DESeq2.beta.prior = TRUE,
  DESeq2.fitType = "parametric",
  DESeq2.normalize = "CPM",
  edgeR.Test = "QLFT",
  edgeR.normalize = "TMM",
  limma.method.fit = "ls",
  limma.trend = TRUE,
  limma.robust = TRUE,
  limma.normalize = "CPM",
  Seurat.normalize = "CPM",
  Seurat.method = "bimod",
  MAST.method = "bayesglm",
  MAST.normalize = "CPM",
  MAST.parallel = TRUE,
  monocle.cores = 1,
  monocle.normalize = "CPM",
  scDD.alpha1 = 0.01,
  scDD.mu0 = 0,
  scDD.s0 = 0.01,
  scDD.a0 = 0.01,
  scDD.b0 = 0.01,
  scDD.normalize = "CPM",
  scDD.permutation = 0,
  Ttest.normalize = "CPM",
  Wilcoxon.normalize = "CPM",
  zingeR.edgeR.normalize = "CPM",
  zingeR.edgeR.maxit.EM = 100
)
}
\arguments{
\item{raw.count}{single-cell RNA-seq matrix. The format could be raw read count or normalized matrix. The rows correspond to genes and the columns.}

\item{cell.label}{cell labels information. The cells need be divided into two categories.}

\item{is.normalized}{a boolean variable that defines whether the input raw.count has been normalized? Default is FALSE.}

\item{verbose}{a boolean variable that defines whether to save the DE analysis results and name "Results_DE_individual.RData" in the current working directory.}

\item{BPSC}{a boolean variable that defines whether to perform DE analysis using the BPSC method. Default is TRUE.}

\item{DEsingle}{a boolean variable that defines whether to perform DE analysis using the DEsingle method. Default is TRUE.}

\item{DESeq2}{a boolean variable that defines whether to perform DE analysis using the DESeq2 method. Default is TRUE.}

\item{edgeR}{a boolean variable that defines whether to perform DE analysis using the edgeR method. Default is TRUE.}

\item{MAST}{a boolean variable that defines whether to perform DE analysis using the MAST method. Default is TRUE.}

\item{monocle}{a boolean variable that defines whether to perform DE analysis using the MONOCLE method. Default is TRUE.}

\item{scDD}{a boolean variable that defines whether to perform DE analysis using the scDD method. Default is TRUE.}

\item{Ttest}{a boolean variable that defines whether to perform DE analysis using the T-test method. Default is TRUE.}

\item{Wilcoxon}{a boolean variable that defines whether to perform DE analysis using the Wilcoxon method. Default is TRUE.}

\item{limma}{a boolean variable that defines whether to perform DE analysis using the limma method. Default is TRUE.}

\item{Seurat}{a boolean variable that defines whether to perform DE analysis using the Seurat method. Default is TRUE.}

\item{zingeR.edgeR}{a boolean variable that defines whether to perform DE analysis using the zingeR.edgeR method. Default is TRUE.}

\item{BPSC.coef}{an integer to point out the column index corresponding to the coefficient for the generalized linear mode (GLM) testing in BPSC. Default value is 2.}

\item{BPSC.normalize}{a string variable specifying the type of size factor estimation in BPSC method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{BPSC.parallel}{a boolean variable that defines whether to execute parallel computing for BPSC method. Default is TRUE.}

\item{DEsingle.parallel}{a boolean variable that defines whether to execute parallel computing for DEsingle method. Default is TRUE.}

\item{DEsingle.normalize}{a string variable specifying the type of size factor estimation in DEsingle method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{DESeq2.test}{a string variable specifying the type of test the difference in deviance between a full and reduced model formula in DESeq2 method. Possible values: "Wald" or "LRT". The values represent Wald tests or likelihood ratio test. Default is "Wald".}

\item{DESeq2.parallel}{a boolean variable that defines whether to execute parallel computing for DESeq2 method. Default is TRUE. The parallel computing may fail on Windows system. Default is TRUE.}

\item{DESeq2.beta.prior}{a boolean variable that defines whether or not to put a zero-mean normal prior on the non-intercept coefficient in DESeq2 method. Default is TRUE.}

\item{DESeq2.fitType}{a string variable specifying the type of fitting of dispersions to the mean intensity in DESeq2 method. Possible values: "parametric", "local", "mean". Default is "parametric".}

\item{DESeq2.normalize}{a string variable specifying the type of size factor estimation in DESeq2 method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{edgeR.Test}{a string variable specifying the type of fitting distribution to count data for each gene. Possible values: "LRT", "QLFT". The values represent negative binomial generalized log-linear model and quasi-likelihood negative binomial generalized log-linear model. Default is "QLFT".}

\item{edgeR.normalize}{a string variable specifying the type of size factor estimation in edgeR method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{limma.method.fit}{a string variable specifying the type of fitting method in limma method. Possible values: "ls", "robust". The values represent least squares and robust regression. Default is "ls".}

\item{limma.trend}{a boolean variable that defines whether or not to allow an intensity-trend for the prior variance in limma method. Default is TRUE.}

\item{limma.robust}{a boolean variable that defines whether or not to estimate defined prior information and variance prior against outlier sample variances in limma method. Default is TRUE.}

\item{limma.normalize}{a string variable specifying the type of size factor estimation in limma method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{Seurat.normalize}{a string variable specifying the type of size factor estimation in Seurat method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{Seurat.method}{a string variable specifying the type of test method in Seurat method. Possible values: "LR", "bimod", "roc". The values represent likelihood-ratio test, negative binomial generalized linear model, ROC analysis. Default is "bimod".}

\item{MAST.method}{a string variable specifying the type of test method in MAST method. Possible values: "glm", "glmer", "bayesglm". Default is "bayesglm".}

\item{MAST.normalize}{a string variable specifying the type of size factor estimation in Seurat method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{MAST.parallel}{a boolean variable that defines whether to execute parallel computing for MAST method. Default is TRUE.}

\item{monocle.cores}{the number of cores to be used while testing each gene for differential expression.. Default is 1.}

\item{monocle.normalize}{a string variable specifying the type of size factor estimation in monocle method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{scDD.alpha1}{prior parameter value to be used to model each gene as a mixture of DP normals in scDD method. Default is 0.01.}

\item{scDD.mu0}{prior parameter values to be used to model each gene as a mixture of DP normals in scDD method. Default is 0.}

\item{scDD.s0}{prior parameter values to be used to model each gene as a mixture of DP normals in scDD method. Default is 0.01.}

\item{scDD.a0}{prior parameter values to be used to model each gene as a mixture of DP normals in scDD method. Default is 0.01.}

\item{scDD.b0}{prior parameter values to be used to model each gene as a mixture of DP normals in scDD method. Default is 0.01.}

\item{scDD.normalize}{a string variable specifying the type of size factor estimation in scDD method. Possible values:  "TMM", "RLE", "CPM". Default is "CPM".}

\item{scDD.permutation}{the number of permutations to be used in calculating empirical p-values in scDD method. If the parameter value is set to 0, the full Bayes Factor will not be performed. Else, scDD method takes the nonparametric Kolmogorove-Smirnov test to identify DGEs. Default is 0.}

\item{Ttest.normalize}{a string variable specifying the type of size factor estimation in t-test method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{Wilcoxon.normalize}{a string variable specifying the type of size factor estimation in Wilcoxon method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{zingeR.edgeR.normalize}{a string variable specifying the type of size factor estimation in zingeR.edgeR method. Possible values: "TMM", "RLE", "CPM". Default is "CPM".}

\item{zingeR.edgeR.maxit.EM}{The number of iterations for EM-algorithm in zingeR.edgeR method. If the EM-algorithm does not stop automatically, then, the algorithm may not be convergence. The user need set a larger value. Default is 100.}
}
\value{
a p-values matrix contains the p-values of each differential expression anlysis methods.
}
\description{
This function is implemented to perform individual DE analysis methods.  The current implementation of
scDEA integrates tweleve state-of-the-art methods:  Beta-poisson mixture model (BPSC),
DEsingle, DESeq2, edgeR, Model-based analysis of single-cell transcriptomics (MAST), monocle, scDD,
T-test, Wilcoxon rank sum test (Wilcoxon test), limma, Seurat and zingeR.edgeR.
This function depends on the follwing R package: BPSC, DEsingle, DESeq2, edgeR, MAST, monocle, scDD,
limma, Seurat, zingeR, SingleCellExperiment, dplyr.
These packages will be automatically installed along
with scDEA. \cr
}
\examples{
data("Grun.counts.hvg")
data("Grun.group.information")
# scDD is very slow
Pvals <- scDEA_individual_methods(raw.count = Grun.counts.hvg,
cell.label = Grun.group.information,  verbose = FALSE)
combination.Pvals <- lancaster.combination(Pvals, weight = TRUE, trimmed = 0.2)
adjusted.Pvals <- scDEA.p.adjust(combination.Pvals, adjusted.method = "bonferroni")
}
\author{
Huisheng, Li,  <lihs@mails.ccnu.edu.cn>
}
